refactor(titlebar): dedupe project pill + dot derivation
Extract ProjectPill (dot + project name button + slash) shared by the
agent capsule and the draft capsule, and agentDotStatus so the capsule
and the project-page crumb derive the dot from store state the same
way. Drop the stale position:relative on .ws-cap-wrap — the popover
anchors to .ws-cap-main now.
